feat: Aplicando regra de negócio ao useCase de delação de uma lista
Verificando se ele existe antes de realizar quaisquer modificações

+++ src/application/use_cases/todo/delete-user-todo.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { IDeleteUserTodoRequest } from "./interfaces/delete-user-todo-request";
 import { TodoRepository } from "src/application/repositories/interfaces/todo-repository";
+import { TodoNotFound } from "./errors/todo-not-found";
 
 @Injectable()
 export class DeleteUserTodoUseCase {
@@ -12,6 +13,14 @@ export class DeleteUserTodoUseCase {
     async execute(request: IDeleteUserTodoRequest){
 
         const { todoId, userId } = request;
+
+        const currentTodo = await this.todosRepository.findTodo(todoId);
+
+        if( !currentTodo ){
+
+            throw new TodoNotFound();
+            
+        }
 
         await this.todosRepository.delete(todoId,userId);
